Fix ActivePowerSchemeChanged event
The event was triggered when the power scheme personality changed. But
that is not what we want. There are only three fixed personalities
(typical-, min-, max-power saving). It kinda worked, because the three
personalities have the same GUIDs as the three default power schemes,
and are mapped to those appropriately. Custom created power schemes also
have one of these. Thus, changing between power schemes that have the
same personality would not trigger the event. Now we got the
GUID_ACTIVE_POWERSCHEME setting to trigger our callback. This will be
fired for every change of the active power scheme.