Fix NullReferenceException
When getting the MainModule from the process, we expected it to be
available. Since a valid process needs to have a valid MainModule. But
apparently there are situation in which a process does not have a
MainModule. Maybe in the moment the process gets terminated. Regardless,
we now assume that the process might not have a MainModule at all. And
just assume an empty FileName.