Fix upgrade on install of new version
The LaunchProduct custom action was apparently executed during the
update process, using the installer of the previous version. This seemed
to fail. Now this fix makes it so that even though the custom action is
executed, it doesn't fail the update process. Though it only works for
the following version since the installer of the previous version is
used during the update process.