Fix race condition when showing toast notification
Since the variable, holding the object of ToastDlg, is reused and we
close currently displayed instances, we can run into a race condition.
When two calls to the ShowToastNotification(Guid, string) function try
to close the current instance. The first triggers the closing. The
threat of ToastDlg starts disposing the object. Then the second tries to
trigger the closing again. A try-catch, catching the
ObjectDisposedException should handle it. If the second call can't
trigger the closing, it doesn't have to do anything anyways an can just
wait for it's turn.