Fix display of Active flag of PowerRule
The Active flag of PowerRules could be true even though the rule wasn't
applied. This could happen when the rule was applied previously while
the PowerRules were saved and the program exited. When starting, it
would load the Active flag and assume it was still active even though
the rule wasn't even applied anymore. Now we just don't save the Active
flag to disk anymore.