Fix showing notification without actual change
When another rule is activated that has the same power scheme as the one
already active, be it from another rule or the baseline, we don't want
to show the toast notification for it anymore. Same when no rule applies
but the baselinePowerScheme is the same as the one that is already
active.