Add AppContext
The AppContext allows us to explicitly sync with the UI thread. This
allows us to create and manage the ToastDlg from the UI thread. We can
use Show() instead of ShowDialog(). So we don't have to wait for the
dialog to close anymore. Additionally we reuse the dialog when a new
toast notification overwrites the current one. Avoid any kind of
flickering. All this avoid irregular issues with the dialog going into
the background sometimes.