Fix process cache
We didn't cache processes that caused an exception on creation. Which
seems logical. But the main idea of the cache is to avoid time consuming
creation. Exception on creation, like trying to access a process of
another user without privileges, are time consuming though. Previously,
we didn't get near causing those exception on creating. Now we do, since
we don't store the Process object anymore. So we cache processes that
caused exception by storing a null. Works the same as before.