Fix handling of process termination in RuleManager
The process termination seems to be quite a bit more complex. If a rule
is already applied that has a higher priority, we don't need to apply
another Rule or fallback to baseline. Same if the applied rule is
affected by the termination remains active (had a ActivationCount higher
than 1). If the applied rule is affected by the termination and doesn't
remain active, we have to apply the next highest active rule or fallback
to baseline. I'm pretty confident this works now. But it could sure be
refactored to improve readability. I hope the test cases cover all
perceivable situations.