Fix RuleManager applying baseline to early
When the RuleManager engine is started with PowerLineRules that aren't
activated, the baseline would be applied. This caused unexpected
behavior and was completely unnecessary. Fixed it by moving the
assignment from the ctor to the end of the StartEngine(...) function.
Including the subscription to the ActivePowerSchemeChanged event. When
the engine isn't running, we don't need that event, nor the baseline.