Reduce WMI interval to 1 second
The previous 5 seconds could cause missed events. With 1 second, it's
not perfect, but hopefully good enough at least for now. Alternative
would be ETW but it needs admin privileges to run. So we would need a
service running that forwards the events. Nice about ETW is that events
are forced to be in order. Which is not the case for WMI. Though it's
very unlikely to ever be an issue here.