Replace WMI with manual check of running processes
WMI isn't reliable. It can happen that it won't notify about a creation
or termination of a process. Due to the nature of how it's implemented.
The ideal alternative would be ETW. But that requires admin privileges.
Since this is not really feasible, we're going back to the previous
implementation of iterating over all currently running processes. Now we
check what processes exist now that haven't existed in the previous
iteration and which do not exist anymore.