Add extended logging
By default only regular logging is active. Extended logging logs a lot
more, including process information of all processes on the system.
Since setting changes are also logged, we had to make sure that the
icons of power schemes are stored deterministic. We use SkiaSharp to do
so. We also provide an option for the user to export logs into a
encrypted 7z file.